NestedLayouts: Added the tutorial files.
Used the Nested layouts technique to break down
the design of the web components and then use
HTML and CSS to structure and design them.

The main tag used in this technique is the <div>
tag since it breaks the structure into a number
of block that are easier to style and order.